rcar-gen2: fdpm-kernel-module: Add condition for cleansstate
	Check if kernel source exists before cleanning it

+++ meta-rcar-gen2/recipes-kernel/fdpm-module/fdpm-kernel-module.bb
@@ -51,11 +51,13 @@ do_cleansstate_prepend() {
 }
 
 do_clean_source() {
-    cd ${KERNELSRC}/include/linux/
-    rm -f fdpm_drv.h ${FDPM_INATALL_HEADERS}
+	if [ -d ${KERNELSRC} ] ; then
+		cd ${KERNELSRC}/include/linux/
+		rm -f fdpm_drv.h ${FDPM_INATALL_HEADERS}
 
-    cd  ${KERNELSRC}/include/
-    rm -f fdpm.symvers ${FDPM_INATALL_HEADERS}
+		cd  ${KERNELSRC}/include/
+		rm -f fdpm.symvers ${FDPM_INATALL_HEADERS}
+	fi
 }
 
 PACKAGES = " \
